shift around and wire up example timeline
